webpack to separate folder
