+++ package.json
@@ -5,11 +5,10 @@
   "private": true,
   "main": "server/index.js",
   "scripts": {
-    "start:client": "cross-env NODE_ENV=development webpack serve --open",
+    "start:client": "cross-env NODE_ENV=development webpack serve --open --config ./webpack/",
     "start:server": "cross-env NODE_ENV=development node ./server",
-    "build": "cross-env NODE_ENV=production webpack",
-    "build:server": "webpack --config ./webpack.server.js",
-    "build:dev": "cross-env NODE_ENV=development webpack",
+    "build:client": "cross-env NODE_ENV=production webpack --config ./webpack/",
+    "build:server": "cross-env NODE_ENV=production webpack --config ./webpack/",
     "test": "jest",
     "prettify": "npx prettier --write .",
     "eslint-check": "node ./node_modules/eslint/bin/eslint ./src/**/*.js --fix",
